Migrate wait_for_nav test to use local files
