+++ tests/test_html/redirect_test.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <meta charset="UTF-8" http-equiv="refresh" content="5;url=https://www.wikipedia.org">
+    <meta charset="UTF-8" http-equiv="refresh" content="5;url=sample_page.html">
     <title>This Page will Redirect</title>
 </head>
 <body>
